+++ src/actions/app.js
@@ -2,7 +2,7 @@ import Promise from "promise";
 import { reset } from "redux-form";
 import get from "lodash/get";
 import isEqual from "lodash/isEqual";
-import { onCreateNewUser, onSubmitProposal, onChangeUsername, onChangePassword, onFetchProposalComments } from "./api";
+import { onSubmitProposal, onChangeUsername, onChangePassword, onFetchProposalComments } from "./api";
 import { onFetchProposal as onFetchProposalApi, onSubmitComment as onSubmitCommentApi } from "./api";
 import { onFetchUsernamesById as onFetchUsernamesByIdApi } from "./api";
 import * as sel from "../selectors";
@@ -16,11 +16,6 @@ export const SET_REPLY_PARENT = "SET_REPLY_PARENT";
 export const onRouteChange = () => dispatch => {
   dispatch(cleanErrors());
   dispatch(act.CLEAN_SLATE());
-};
-
-export const onSignup = act.REQUEST_SIGNUP_CONFIRMATION;
-export const onSignupConfirm = props => dispatch => {
-  dispatch(onCreateNewUser(props));
 };
 
 export const onSetReplyParent = (parentId = TOP_LEVEL_COMMENT_PARENTID) =>
